+++ unikmer/cmd/inter.go
@@ -327,5 +327,5 @@ func init() {
 	RootCmd.AddCommand(interCmd)
 
 	interCmd.Flags().StringP("out-prefix", "o", "-", `out file prefix ("-" for stdout)`)
-	interCmd.Flags().BoolP("sort", "s", false, "sort Kmers, this reduces file size, you can even disable gzip compression by flag -C/--no-compress")
+	interCmd.Flags().BoolP("sort", "s", false, helpSort)
 }